Store locally the modifications/deletions done by the users
By saving them in the local storage and retrieving them when `<App />` mount before trying to fetch from the remote endpoint.

Note that we naively assume that the endpoint data is not going to be changed, so we took the local modified ones as the "good ones", otherwise would have to make the logic a bit different (making a diff from the users of both places and merge them/decide which ones to take...etc).

Also, we call `saveUsersLocally(newUsers)` when editing & deleting.  On a more complex app, we could put this inside an `useEffect` that triggers when `users` is modified, but for now we do it like this.